src/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
apyori==1.1.2
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
exceptiongroup==1.3.0
fastapi==0.117.1
gensim==4.3.3
greenlet==3.2.4
h11==0.16.0
idna==3.10
joblib==1.5.2
nltk==3.9.1
numpy==1.26.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pypdfium2==4.30.0
python-dotenv==1.1.1
python-multipart==0.0.20
regex==2025.9.18
scikit-learn==1.7.2
scipy==1.13.1
setuptools==80.9.0
smart_open==7.3.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
wheel==0.45.1
wrapt==1.17.3
